requirements.txt
pandas>=2.1.0
pyarrow>=14.0.0
pandera>=0.20.0
numpy<2
torch>=2.3.0
faiss-cpu==1.8.0
scikit-learn>=1.4.0
tqdm>=4.66.0
